requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3